+++ lib/src/modules/engineer/view_reports/parts/model/engineer_part_report_response.g.dart
@@ -17,6 +17,7 @@ EngineerPartReportResponse _$EngineerPartReportResponseFromJson(
       ..refId = json['r_id'] as String?
       ..isSuccess = json['s'] as bool? ?? false
       ..errorMsg = json['em'] as String?
+      ..successMessage = json['sm'] as String?
       ..data = json['dt'] == null
           ? null
           : PartReportData.fromJson(json['dt'] as Map<String, dynamic>);
@@ -29,5 +30,6 @@ Map<String, dynamic> _$EngineerPartReportResponseToJson(
       'r_id': instance.refId,
       's': instance.isSuccess,
       'em': instance.errorMsg,
+      'sm': instance.successMessage,
       'dt': instance.data,
     };