+++ lib/src/modules/trc_executive/models/device_receive_response.g.dart
@@ -17,6 +17,7 @@ DeviceReceiveResponse _$DeviceReceiveResponseFromJson(
       ..refId = json['r_id'] as String?
       ..isSuccess = json['s'] as bool? ?? false
       ..errorMsg = json['em'] as String?
+      ..successMessage = json['sm'] as String?
       ..data = json['dt'] == null
           ? null
           : DeviceReceiveData.fromJson(json['dt'] as Map<String, dynamic>);
@@ -29,6 +30,7 @@ Map<String, dynamic> _$DeviceReceiveResponseToJson(
       'r_id': instance.refId,
       's': instance.isSuccess,
       'em': instance.errorMsg,
+      'sm': instance.successMessage,
       'dt': instance.data,
     };